+++ src/App.js
@@ -1,12 +1,17 @@
 import './App.scss';
 
-import Header from './components/Header/Header'
-import Routes from './components/Routes'
 import { useCookies } from 'react-cookie';
 import { BrowserRouter } from 'react-router-dom';
 import { useSelector } from 'react-redux';
+import { ToastContainer } from 'react-toastify';
+
+import { SettingsContext } from './settings';
+import Header from './components/Header/Header'
+import Routes from './components/Routes'
 import GlobalNav from './components/GlobalNav/GlobalNav';
 import Footer from './components/Footer/Footer';
+import BetaBanner from './components/BetaBanner/BetaBanner';
+import BetaModal from './components/Modals/BetaModal';
 
 function App() {
   const isEmbedded = useSelector((state) => state.editor.isEmbedded);
@@ -16,11 +21,16 @@ function App() {
     <div 
     id='app'
     className = {`--${cookies.theme || themeDefault } font-size-${cookies.fontSize || 'small' }`}>
-      <BrowserRouter>
-        { isEmbedded ? null : <><GlobalNav/><Header/></> }
-        <Routes />
-        { isEmbedded ? null : <Footer/> }
-      </BrowserRouter>
+      
+      <SettingsContext.Provider value={{theme: cookies.theme || themeDefault, fontSize: cookies.fontSize || 'small' }}>
+        <BrowserRouter>
+          { isEmbedded ? null : <><GlobalNav/><BetaBanner/><Header/></> }
+          <Routes />
+          { isEmbedded ? null : <Footer/> }
+          <BetaModal/>
+        </BrowserRouter>
+        <ToastContainer position='bottom-center' className='toast--bottom-center' />
+      </SettingsContext.Provider>
     </div>
   );
 }