+++ src/components/FilePane/ProjectImages/ProjectImages.js
@@ -1,15 +1,17 @@
 import './ProjectImages.scss';
 
 import { useSelector } from 'react-redux'
-import { ChevronDown } from '../../../../Icons';
+import { ChevronDown } from '../../../Icons';
+import { useTranslation } from 'react-i18next'
 
 const ProjectImages = () => {
   const projectImages = useSelector((state) => state.editor.project.image_list);
+  const { t } = useTranslation()
 
   return (
-    <details className='file-menu-section file-menu-section__images' open>
+    <details className='file-pane-section file-pane-section__images' open>
       <summary>
-        <h2 className='menu-pop-out-subheading'>Image Gallery</h2>
+        <h2 className='menu-pop-out-subheading'>{t('filePane.images')}</h2>
         <div className='accordion-icon'>
           <ChevronDown />
         </div>