+++ src/components/Footer/Footer.test.js
@@ -5,25 +5,25 @@ import Footer from "./Footer";
 
 test('Footer renders', () => {
   render(<Footer/>)
-  expect(screen.queryByText(/Raspberry Pi Foundation/)).toBeInTheDocument()
+  expect(screen.queryByText('footer.charityNameAndNumber')).toBeInTheDocument()
 })
 
 test('Links to privacy policy', () => {
   render(<Footer/>)
-  expect(screen.queryByText('Privacy')).toHaveAttribute('href', 'https://www.raspberrypi.org/privacy')
+  expect(screen.queryByText('footer.privacy')).toHaveAttribute('href', 'https://www.raspberrypi.org/privacy')
 })
 
 test('Links to cookie policy', () => {
   render(<Footer/>)
-  expect(screen.queryByText('Cookies')).toHaveAttribute('href', 'https://www.raspberrypi.org/cookies')
+  expect(screen.queryByText('footer.cookies')).toHaveAttribute('href', 'https://www.raspberrypi.org/cookies')
 })
 
 test('Links to accessibility policy', () => {
   render(<Footer/>)
-  expect(screen.queryByText('Accessibility')).toHaveAttribute('href', 'https://www.raspberrypi.org/accessibility')
+  expect(screen.queryByText('footer.accessibility')).toHaveAttribute('href', 'https://www.raspberrypi.org/accessibility')
 })
 
 test('Links to safeguarding policy', () => {
   render(<Footer/>)
-  expect(screen.queryByText('Safeguarding')).toHaveAttribute('href', 'https://www.raspberrypi.org/safeguarding')
+  expect(screen.queryByText('footer.safeguarding')).toHaveAttribute('href', 'https://www.raspberrypi.org/safeguarding')
 })