+++ src/components/Header/Header.js
@@ -1,6 +1,6 @@
 import './Header.scss'
 import { useSelector, connect, useDispatch } from 'react-redux'
-import { toast } from 'react-toastify';
+import { useTranslation } from 'react-i18next';
 import Button from '../Button/Button';
 import { SettingsIcon, SquaresIcon } from '../../Icons';
 import { saveProject, updateProject } from '../../utils/apiCallHandler';
@@ -11,56 +11,62 @@ import SettingsMenu from '../Menus/SettingsMenu/SettingsMenu';
 import ProjectName from './ProjectName';
 
 import editor_logo from '../../assets/editor_logo.svg'
+import DownloadButton from './DownloadButton';
+import { showSavedMessage } from '../../utils/Notifications';
 
 
 const Header = (props) => {
   const { user } = props;
   const project = useSelector((state) => state.editor.project);
+  const projectLoaded = useSelector((state) => state.editor.projectLoaded)
 
   const dispatch = useDispatch();
   let history = useHistory();
+  const { t } = useTranslation()
 
   const onClickSave = async () => {
     window.plausible('Save button')
 
     if (!project.identifier) {
       const response = await saveProject(project, user.access_token)
-      const identifier = response.data.identifier;
-      const project_type = response.data.project_type;
-      dispatch(setProjectLoaded(false));
-      history.push(`/${project_type}/${identifier}`)
-      return;
+
+      if (response.status === 200) {
+        const identifier = response.data.identifier;
+        const project_type = response.data.project_type;
+        dispatch(setProjectLoaded(false));
+        history.push(`/${project_type}/${identifier}`)
+        showSavedMessage()
+      }
     }
+    else {
+      const response = await updateProject(project, user.access_token)
 
-    const response = await updateProject(project, user.access_token)
-
-    if(response.status === 200) {
-      dispatch(setProject(response.data));
-      toast("Project saved!", {
-        position: toast.POSITION.TOP_CENTER
-      });
+      if(response.status === 200) {
+        dispatch(setProject(response.data));
+        showSavedMessage()
+      }
     }
   }
 
   return (
     <div className='editor-header-wrapper'>
       <header className='editor-header'>
-        <img className='editor-logo' src={editor_logo} alt='Editor logo'/>
+        <img className='editor-logo' src={editor_logo} alt={t('header.editorLogoAltText')}/>
         { user !== null ? (
           <a href='/projects' className='project-gallery-link'>
             {<><SquaresIcon />
-            <span className='editor-header__text'>My Projects</span></>}</a>
+            <span className='editor-header__text'>{t('header.projects')}</span></>}</a>
         ) : null }
-        <ProjectName />
+        { projectLoaded ? <ProjectName /> : null }
         <div className='editor-header__right'>
+          { projectLoaded ? <DownloadButton /> : null }
           <Dropdown
             ButtonIcon={SettingsIcon}
-            buttonText='Settings'
-            buttonTextClassName='editor-header__text'
+            buttonText={t('header.settings')}
             MenuContent={SettingsMenu} />
 
-          {user !== null && project.user_id === user.profile.user ? (
-            <Button className='btn--save' onClickHandler = {onClickSave} buttonText = "Save" />
+          {projectLoaded && user !== null && (project.user_id === user.profile.user || !project.identifier) ? (
+            <Button className='btn--save' onClickHandler = {onClickSave} buttonText = {t('header.save')} />
           ) : null }
         </div>
       </header>