+++ src/components/Header/Header.test.js
@@ -1,8 +1,9 @@
 import React from "react";
-import { fireEvent, render } from "@testing-library/react"
+import { fireEvent, render, screen, waitFor } from "@testing-library/react"
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
 import axios from "axios";
+import { showSavedMessage } from "../../utils/Notifications";
 
 import Header from "./Header";
 
@@ -15,11 +16,12 @@ jest.mock('react-router-dom', () => ({
   })
 }));
 
+jest.mock('../../utils/Notifications')
+
+
 describe("When logged in and user owns project", () => {
   let store;
   let saveButton;
-  let getByText;
-  let queryByText;
 
   beforeEach(() => {
     const middlewares = []
@@ -32,6 +34,7 @@ describe("When logged in and user owns project", () => {
           image_list: [],
           user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
         },
+        projectLoaded: true,
       },
       auth: {
         user: {
@@ -43,12 +46,12 @@ describe("When logged in and user owns project", () => {
       }
     }
     store = mockStore(initialState);
-    ({getByText, queryByText} = render(<Provider store={store}><Header/></Provider>));
-    saveButton = getByText(/Save/)
+    render(<Provider store={store}><Header/></Provider>);
+    saveButton = screen.queryByText('header.save')
   })
 
   test("Save button renders", () => {
-      expect(saveButton.textContent).toBe("Save");
+      expect(saveButton).toBeInTheDocument();
   })
 
   test("Clicking save button sends PUT request to correct endpoint", () => {
@@ -72,15 +75,28 @@ describe("When logged in and user owns project", () => {
     expect(actions).toEqual([expectedPayload])
   })
 
+  test("Successful save prompts success message", async () => {
+    axios.put.mockImplementationOnce(() => Promise.resolve({ status: 200, data: {}}))
+    fireEvent.click(saveButton)
+    await waitFor(() => expect(showSavedMessage).toHaveBeenCalled())
+  })
+
   test("Renders project gallery link", () => {
-    expect(queryByText("My Projects")).not.toBeNull();
+    expect(screen.queryByText('header.projects')).not.toBeNull();
+  })
+
+  test('Download button shown', () => {
+    expect(screen.queryByText('header.download')).toBeInTheDocument()
+  })
+
+  test('Project name is shown', () => {
+    expect(screen.queryByRole('textbox')).toBeInTheDocument()
   })
 })
 
 describe("When logged in and no project identifier", () => {
   let store;
   let saveButton;
-  let getByText;
 
   beforeEach(() => {
     const middlewares = []
@@ -92,6 +108,7 @@ describe("When logged in and no project identifier", () => {
           image_list: [],
           user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
         },
+        projectLoaded: true,
       },
       auth: {
         user: {
@@ -103,8 +120,16 @@ describe("When logged in and no project identifier", () => {
       }
     }
     store = mockStore(initialState);
-    ({getByText} = render(<Provider store={store}><Header/></Provider>));
-    saveButton = getByText(/Save/)
+    render(<Provider store={store}><Header/></Provider>);
+    saveButton = screen.getByText('header.save')
+  })
+
+  test('Download button shown', () => {
+    expect(screen.queryByText('header.download')).toBeInTheDocument()
+  })
+
+  test("Save button is shown", () => {
+    expect(saveButton).toBeInTheDocument()
   })
 
   test("Clicking save creates new project", () => {
@@ -118,11 +143,20 @@ describe("When logged in and no project identifier", () => {
     const headers = {"headers": {"Accept": "application/json", "Authorization": access_token}}
     expect(axios.post).toHaveBeenCalledWith(`${api_host}/api/projects`, {"project": new_project}, headers)
   })
+
+  test('Project name is shown', () => {
+    expect(screen.queryByRole('textbox')).toBeInTheDocument()
+   })
+   
+  test("Successful save prompts success message", async () => {
+    axios.post.mockImplementationOnce(() => Promise.resolve({ status: 200, data: {}}))
+    fireEvent.click(saveButton)
+    await waitFor(() => expect(showSavedMessage).toHaveBeenCalled())
+  })
 })
 
 describe("When logged in and user does not own project", () => {
   let store;
-  let queryByText;
 
   beforeEach(() => {
     const middlewares = []
@@ -135,6 +169,7 @@ describe("When logged in and user does not own project", () => {
           image_list: [],
           user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
         },
+        projectLoaded: true,
       },
       auth: {
         user: {
@@ -146,20 +181,27 @@ describe("When logged in and user does not own project", () => {
       }
     }
     store = mockStore(initialState);
-    ({queryByText} = render(<Provider store={store}><Header/></Provider>));
+    render(<Provider store={store}><Header/></Provider>);
   })
 
   test("Renders project gallery link", () => {
-    expect(queryByText("My Projects")).not.toBeNull();
+    expect(screen.queryByText('header.projects')).not.toBeNull();
   })
 
   test("No save button", () => {
-    expect(queryByText("Save")).toBeNull()
+    expect(screen.queryByText('header.save')).toBeNull()
+  })
+
+  test('Download button shown', () => {
+    expect(screen.queryByText('header.download')).toBeInTheDocument()
+  })
+
+  test('Project name is shown', () => {
+    expect(screen.queryByText('header.newProject')).toBeInTheDocument()
   })
 })
 
 describe("When not logged in", () => {
-  let queryByText;
 
   beforeEach(() => {
     const middlewares = []
@@ -171,20 +213,60 @@ describe("When not logged in", () => {
             components: [],
             image_list: [],
           },
+          projectLoaded: true,
         },
         auth: {
           user: null
         }
       }
     const store = mockStore(initialState);
-    ({queryByText} = render(<Provider store={store}><Header/></Provider>));
+    render(<Provider store={store}><Header/></Provider>);
   })
 
   test("No save button", () =>{
-    expect(queryByText("Save")).toBeNull();
+    expect(screen.queryByText('header.save')).toBeNull();
   })
 
   test("No project gallery link", () => {
-    expect(queryByText("My Projects")).toBeNull();
+    expect(screen.queryByText('header.projects')).toBeNull();
+  })
+
+  test('Download button shown', () => {
+    expect(screen.queryByText('header.download')).toBeInTheDocument()
+  })
+
+  test('Project name is shown', () => {
+    expect(screen.queryByText('header.newProject')).toBeInTheDocument()
+  })
+})
+
+describe('When no project loaded', () => {
+
+  beforeEach(() => {
+    const middlewares = []
+    const mockStore = configureStore(middlewares)
+    const initialState = {
+        editor: {
+          project: {},
+          projectLoaded: false,
+        },
+        auth: {
+          user: "5254370e-26d2-4c8a-9526-8dbafea43aa9"
+        }
+      }
+    const store = mockStore(initialState);
+    render(<Provider store={store}><Header/></Provider>);
+  })
+
+  test('No project name', () => {
+    expect(screen.queryByText('header.newProject')).not.toBeInTheDocument()
+  })
+
+  test('No download button', () => {
+    expect(screen.queryByText('header.download')).not.toBeInTheDocument()
+  })
+
+  test('No save button', () => {
+    expect(screen.queryByText('header.save')).not.toBeInTheDocument()
   })
 })