+++ src/setupTests.js
@@ -19,3 +19,19 @@ window.matchMedia = (query) => ({
 })
 
 window.plausible = jest.fn()
+
+jest.mock('react-i18next', () => ({
+  // this mock makes sure any components using the translate hook can use it without a warning being shown
+  useTranslation: () => {
+    return {
+      t: (str) => str,
+      i18n: {
+        changeLanguage: () => new Promise(() => {}),
+      },
+    };
+  },
+}));
+
+jest.mock('./i18n', () => ({
+  t: (string) => string
+}))